+++ src/mesh/primitive/Cylinder.cc
@@ -24,9 +24,9 @@ using namespace std;
 
 shared_ptr<Mesh> Cylinder::Mesh(float radius,
 								float height,
-								int slices,
-								int segments,
-								int rings,
+								unsigned slices,
+								unsigned segments,
+								unsigned rings,
 								const shared_ptr<Material> material) {
 
 	return make_shared<a3d::Mesh>("Cylinder",
@@ -44,9 +44,9 @@ shared_ptr<Mesh> Cylinder::Mesh(float radius,
 
 Cylinder::Cylinder(float radius,
 				   float height,
-				   int slices,
-				   int segments,
-				   int rings):
+				   unsigned slices,
+				   unsigned segments,
+				   unsigned rings):
 		MeshElement{},
 		_radius{radius},
 		_height{height},
@@ -61,7 +61,7 @@ Cylinder::Cylinder(float radius,
 	/// @param start Counterclockwise angle around the z-axis relative to the x-axis.
 	/// @param sweep Counterclockwise angle around the z-axis.
 
-	auto cylinder = CappedCylinderMesh{radius, height/2.0, slices, segments, rings};
+	auto cylinder = CappedCylinderMesh{radius, height/2.0, (int)slices, (int)segments, (int)rings};
 
 	for (const MeshVertex& v : cylinder.vertices()) {
 		_vertices.push_back({ vec3(v.position[0], v.position[1], v.position[2]),
@@ -92,14 +92,14 @@ float Cylinder::height() const {
 	return _height;
 }
 
-int Cylinder::slices() const {
+unsigned Cylinder::slices() const {
 	return _slices;
 }
 
-int Cylinder::segments() const {
+unsigned Cylinder::segments() const {
 	return _segments;
 }
 
-int Cylinder::rings() const {
+unsigned Cylinder::rings() const {
 	return _rings;
 }
